SisDis/requirements.txt
kafka-python
pymongo
pandas
matplotlib
scikit-learn
apache-flink